requirements.txt
Flask
pdf2image
Pillow
pytesseract
